refactor(mvm-build): migrate NixStoreImageLock into BuilderVmRuntime
Plan 97 Phase C, fourth migration after stage_job_dir (#434),
read_job_result (#435), and finalize_flake_job/finalize_install_job
(#436). Lifts the host-side nix-store image lock out of libkrun_builder.rs
into builder_vm_runtime.rs so the future VzBuilderVm path reuses it.

Migrated:
- NixStoreImageLock struct (pub, with pub path() accessor)
- acquire_nix_store_image_lock function (pub)
- 4 unit tests (2 existing + 2 new)

The new signature takes builder_cache_dir: &Path as the first arg so
the helper stays free of mvm_core::config lookups — callers compose
the path themselves (libkrun_builder.rs uses its existing
builder_vm_cache_dir() helper). This also removes the XDG_CACHE_HOME
env-var hack the libkrun-side tests needed; each migrated test now
passes a fresh TempDir path directly and runs without process-wide env
mutation. Two new tests pin the filename convention
(nix-store-<arch>.img) and the create-missing-cache-dir contract.

NixStoreImageLock's _file: std::fs::File is load-bearing — the type
docs call this out explicitly per Plan 97 §"Phase C seam design" so
underscore-prefixed reads as inert don't mislead future readers.
